Moved core/db to internal/pkg/db
