+++ core/data/event.go
@@ -22,8 +22,8 @@ import (
 	"time"
 
 	"github.com/edgexfoundry/edgex-go/core/clients/types"
-	"github.com/edgexfoundry/edgex-go/core/db"
 	"github.com/edgexfoundry/edgex-go/core/domain/models"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db"
 	"github.com/gorilla/mux"
 )
