+++ core/data/init.go
@@ -21,11 +21,11 @@ import (
 	"github.com/edgexfoundry/edgex-go/core/clients/types"
 	"github.com/edgexfoundry/edgex-go/core/data/interfaces"
 	"github.com/edgexfoundry/edgex-go/core/data/messaging"
-	"github.com/edgexfoundry/edgex-go/core/db"
-	"github.com/edgexfoundry/edgex-go/core/db/influx"
-	"github.com/edgexfoundry/edgex-go/core/db/memory"
-	"github.com/edgexfoundry/edgex-go/core/db/mongo"
 	"github.com/edgexfoundry/edgex-go/internal"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db/influx"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db/memory"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db/mongo"
 	consulclient "github.com/edgexfoundry/edgex-go/support/consul-client"
 	"github.com/edgexfoundry/edgex-go/support/logging-client"
 )
