+++ core/data/valuedescriptor.go
@@ -23,8 +23,8 @@ import (
 	"fmt"
 
 	"github.com/edgexfoundry/edgex-go/core/clients/types"
-	"github.com/edgexfoundry/edgex-go/core/db"
 	"github.com/edgexfoundry/edgex-go/core/domain/models"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db"
 	"github.com/gorilla/mux"
 )
