+++ core/metadata/init.go
@@ -17,11 +17,11 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/edgexfoundry/edgex-go/core/db"
-	"github.com/edgexfoundry/edgex-go/core/db/memory"
-	"github.com/edgexfoundry/edgex-go/core/db/mongo"
 	"github.com/edgexfoundry/edgex-go/core/metadata/interfaces"
 	"github.com/edgexfoundry/edgex-go/internal"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db/memory"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db/mongo"
 	consulclient "github.com/edgexfoundry/edgex-go/support/consul-client"
 	logger "github.com/edgexfoundry/edgex-go/support/logging-client"
 	notifications "github.com/edgexfoundry/edgex-go/support/notifications-client"
