+++ core/metadata/rest_deviceservice.go
@@ -21,8 +21,8 @@ import (
 	"net/url"
 	"strconv"
 
-	"github.com/edgexfoundry/edgex-go/core/db"
 	"github.com/edgexfoundry/edgex-go/core/domain/models"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db"
 	"github.com/gorilla/mux"
 	"gopkg.in/mgo.v2/bson"
 )
