+++ internal/pkg/db/influx/data.go
@@ -20,7 +20,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/edgexfoundry/edgex-go/core/db"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db"
 	"github.com/edgexfoundry/edgex-go/core/domain/models"
 	client "github.com/influxdata/influxdb/client/v2"
 	"gopkg.in/mgo.v2/bson"
