+++ internal/pkg/db/mongo/data.go
@@ -14,7 +14,7 @@
 package mongo
 
 import (
-	"github.com/edgexfoundry/edgex-go/core/db"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db"
 	"github.com/edgexfoundry/edgex-go/core/domain/models"
 	"gopkg.in/mgo.v2/bson"
 )
