+++ internal/pkg/db/mongo/metadata.go
@@ -16,7 +16,7 @@ package mongo
 import (
 	"errors"
 
-	"github.com/edgexfoundry/edgex-go/core/db"
+	"github.com/edgexfoundry/edgex-go/internal/pkg/db"
 	"github.com/edgexfoundry/edgex-go/core/domain/models"
 	"gopkg.in/mgo.v2/bson"
 )
